feat: 절대경로 설정
