+++ src/styles/main.css
@@ -1,0 +1,3 @@
+@import './layout.css';
+@import './post.css';
+@import './writing.css';
